fix user status
