+++ frontend/src/components/organisms/Modals/index.js
@@ -1,0 +1,4 @@
+import UserStatus from './UserStatus'
+
+// eslint-disable-next-line import/prefer-default-export
+export { UserStatus }
