+++ frontend/src/components/organisms/index.js
@@ -1,5 +1,6 @@
 import { BarChart, LineChart, PieChart } from './Charts'
 import { MapOfDistricts } from './InteractiveMap'
+import { UserStatus } from './Modals'
 import Modal from './Modal/Modal'
 import AdCard from './AdCard'
 import Gallery from './Gallery/Gallery'
@@ -9,15 +10,16 @@ import Pagination from './Pagination'
 import ProtectedRoute from './ProtectedRoute'
 
 export {
-  Modal,
-  MapOfDistricts,
-  BarChart,
-  LineChart,
-  PieChart,
   AdCard,
-  Gallery,
+  BarChart,
   ContactModal,
+  Gallery,
+  LineChart,
   Map,
+  MapOfDistricts,
+  Modal,
   Pagination,
+  PieChart,
   ProtectedRoute,
+  UserStatus,
 }
