+++ frontend/src/utils/paths.js
@@ -1,12 +1,13 @@
 const paths = {
-  home: '/',
-  profile: '/profile',
-  bills: '/my-bills',
-  userAds: '/user-ads',
-  newAd: '/new-ad',
   ads: '/ads',
+  bills: '/my-bills',
+  home: '/',
   login: '/login',
+  newAd: '/new-ad',
+  profile: '/profile',
   register: '/register',
+  userAds: '/user-ads',
+  userAdmin: '/lista-usuarios-admins',
 }
 
 export default paths
